Add endpoint, add test, test_get_list_web_api(need help)
